[TR] Update default tags/detail in audit events
Removes all default fields from the detail part of an audit event.
For explicit audit events detail, will now only contain fields which
are specific to the audit event.

deviceID is moved to the tags section as this is still useful for all
audit events (this was already the case for some manually created tags).

OutboundCall audit events still have ipAddress, authorisation and token.

Makes transactionName optional as this does not always provide additional
information as auditType is usually enough.

Many callers do not use 'toAuditDetails' when populating the detail
part of an audit event anyway and these fields (apart from deviceID) are
not very useful.

The intention of this change is to try to get some consistency across
audit events and to remove unnecessary data.