[TR] TXM-3192 Update default tags/detail in audit events
The intention of this change is to try to get some consistency across
audit events and to remove unnecessary data.

Removes all default fields from the detail part of an audit event.
'detail' is now used for data which is specific to the auditType.

The toAuditDetails method has been retained to prevent compilation
failures in the many projects which depend on play-auditing

deviceID is moved to the tags section as this is still useful for all
audit events (this was already the case for some manually created tags).

OutboundCall audit events still have ipAddress, authorisation and token
captured in 'detail'.

This commit also makes transactionName optional. Usually auditType is enough.

For reference, many projects do not use 'toAuditDetails' when populating the detail
part of an audit event anyway.