[TR] remove separate execution context reference
Using the passed in excution context rather than explicitly using the
play execution context so that the dependency on play can be removed.

Also removed the auditFromPlayFrontend method as this is never called.

There is a possible issue with the use of the play thread pool due to
the use of blocking calls when posting to datastream. However, this
commit does not change the execution context used and no issues have
been reported since the play execution context was used in this way.