[TR] TXM-3187 Remove default datastream base uri
The default datastream is not used by any projects as they all
specify the datastream url in application.conf

This change makes the uri required.

It is more involved than expected because in the case where auditing
was disabled the default value was loaded, even though it was not used.